Added a new FSOP for grabbing a comprehensive FilePkg struct, rather than piecemeal information about a file.
Improved the handling of parseFSTreeRec return codes and short-circuiting.
Removed the Inode struct.
Replaced Inode entries in the DirNode struct with FilePkg entries.
Disabled getInode, getName, and dirList until I can recode things without the Inode struct.
Implemented a proper CreateFile/Cleanup persistence system that grabs a FilePkg when a file is opened, stores it in the meantime, and destroys it when the file is closed, obviating the need to thrash the FS tree every single time a file is queried for information within its lifetime.
Updated btrfsGetFileInformation so it uses the new persistent FilePkg array rather than directly querying the FS tree.
Made a few small Dokan callback adjustments involving mutex placement and other things.