Added two new FSOPs to enable buttery-smooth directory listings.
Reimplemented btrfsFindFiles using the new FSOPs; the entire function requires parsing the FS tree only twice!
Directory listings are now O(1) with respect to the number of files in the directory instead of O(n).
Added some more descriptive error messages to a couple of recently-changed Dokan callbacks.
Removed the no-duplicates assumption in btrfsCreateFile and btrfsOpenDirectory to prevent assertion failures until openFiles gets reimplemented as a multimap.
Fixed a major issue with the newer short-circuiting mechanism in the FS tree parser.
Fixed a couple of previously-overlooked endianness conversions in the FS tree parser.
Removed unneeded code for destroying DirList structs.