Created a new FSOP to perform an entire directory listing with a single FS tree walk, enabling the very silky-smoothiest directory listings to date.
Extended parseFSTreeRec to accept a temp parameter, which was needed in the one-pass dirlist for the '..' inode since it actually comes before we know what its object ID is.
Fixed a heap corruption bug involving a double-free upon the failure of an FSOP_DIR_LIST_B operation (was 100% too comprehensive about freeing the struct in question).