Replaced the idiosyncratic device, chunk, and root arrays with vectors containing items from the trees.
Removed device and chunk dumping for now; it will return in the same form as FS tree dumping is now.
Changed the type of FSOP args from int to FSOperation for type safety.
Changed the last member of the BtrfsInodeRef struct into a zero-size char (variably sized struct).
Redefined BtrfsChunkItem as a variably sized struct containing a number of BtrfsChunkItemStripes at the end.
Removed the Chunk struct and introduced the ItemPlus struct, containing a BtrfsItem and data associated with the item.