Finished the modifications to getFileID necessary to accomodate subvolumes.
Replaced the BtrfsObjID stored as Dokan's context variable with a pointer to a FileID struct.
Added another value to DIR_ITEM dumps that indicates whether the particular item represents a subvolume.
Replaced the parentID value in the FilePkg struct (previously a BrtfsObjID) with a FileID struct and modified FS tree parser code to handle the change.
Modified FSOP_NAME_TO_ID so it now outputs whether the ID in question refers to a subvolume.
Added printf statements to all TODO comments to facilitate quicker mitigation of issues.
Modified the program stdout header.
Improved the speed of UUID-to-string conversions and removed braces from the output.
Removed the printf statement after successful SB validation.