Fixed the '.' and '..' issue once and for all, hopefully.
Enhanced getPathID so it now returns the parent file ID.
Modified the failed read assertion in the loadNode function for better clarity.
Modified btrfsCreateFileCommon so it now populates the new parent fields.
Modified FSOP_DIR_LIST so it takes a FilePkg for the parent rather than just a BrtfsObjID.
Added a field in the FilePkg struct for the parent's inode.
Ripped all of the parent-related functionality and '.'/'..'-specific code out of the FS tree parser.
Removed the temp item from the FS tree parsing functions, as finding parent information is now thankfully done in advance.