Reworked argument handling so that options can come anywhere, and multiple devices can be specified.
Replaced the single BlockReader instance with an array of instances, one for each device specified.
Replaced the single devicePath string with a vector of multiple device paths.
Added a warning in each location where device reads occur, indicating that work should be done to ensure that the proper device is chosen to be read from in the future.